Refactor MQTT-SN code
