+++ src/include.am
@@ -8,6 +8,11 @@ src_libwolfmqtt_la_SOURCES = src/mqtt_client.c \
                              src/mqtt_packet.c \
                              src/mqtt_socket.c
 
+if BUILD_SN
+src_libwolfmqtt_la_SOURCES += src/mqtt_sn_client.c \
+                              src/mqtt_sn_packet.c
+endif
+
 src_libwolfmqtt_la_CFLAGS       = -DBUILDING_WOLFMQTT $(AM_CFLAGS)
 src_libwolfmqtt_la_CPPFLAGS     = -DBUILDING_WOLFMQTT $(AM_CPPFLAGS)
 src_libwolfmqtt_la_LDFLAGS      = ${AM_LDFLAGS} -no-undefined -version-info ${WOLFMQTT_LIBRARY_VERSION}
