+++ wolfmqtt/include.am
@@ -11,3 +11,8 @@ nobase_include_HEADERS+= \
                          wolfmqtt/visibility.h \
                          wolfmqtt/options.h \
                          wolfmqtt/vs_settings.h
+
+if BUILD_SN
+nobase_include_HEADERS+= wolfmqtt/mqtt_sn_client.h \
+                         wolfmqtt/mqtt_sn_packet.h
+endif
